+++ django_prose_editor/static/django_prose_editor/init.js
@@ -1,6 +1,26 @@
 /* global DjangoProseEditor */
-document.addEventListener("DOMContentLoaded", () => {
-  for (let el of document.querySelectorAll("[data-django-prose-editor]")) {
-    DjangoProseEditor.createEditor(el)
+
+const marker = "data-django-prose-editor"
+
+function initializeDjangoProseEditor(container) {
+  for (let el of container.querySelectorAll(`[${marker}]`)) {
+    if (!el.id.includes("__prefix__")) {
+      DjangoProseEditor.createEditor(el)
+      el.removeAttribute(marker)
+    }
   }
+}
+
+function initializeDjangoInlines() {
+  let o
+  if ((o = window.django) && (o = o.jQuery)) {
+    o(document).on("formset:added", (e) => {
+      initializeDjangoProseEditor(e.target)
+    })
+  }
+}
+
+document.addEventListener("DOMContentLoaded", () => {
+  initializeDjangoProseEditor(document)
+  initializeDjangoInlines()
 })