+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Change log
 Next version
 ~~~~~~~~~~~~
 
+- Fixed the initialization in Django admin inlines.
+
 
 `0.1`_ (2024-03-11)
 ~~~~~~~~~~~~~~~~~~~